refactor(catalog): extract import matching and usage
