+++ backend/app/services/component_catalog_domain.py
@@ -48,6 +48,7 @@ from app.services.catalog.normalization import (
     sha256_file as _sha256_file,
 )
 from app.services.catalog.project_import_sessions import CatalogProjectImportSessions
+from app.services.catalog.project_import_matching import CatalogProjectImportMatching
 from app.services.catalog.revision_comparison import CatalogRevisionComparison
 from app.services.catalog.revision_kernel import (
     CatalogRevisionKernel,
@@ -446,6 +447,7 @@ class ComponentCatalogDomainService:
     _component_read_models: CatalogComponentReadModels = CatalogComponentReadModels(_revision_kernel)
     _component_queries: CatalogComponentQueries = CatalogComponentQueries(_component_read_models)
     _project_import_sessions: CatalogProjectImportSessions = CatalogProjectImportSessions()
+    _project_import_matching: CatalogProjectImportMatching = CatalogProjectImportMatching()
 
     def __init__(self, store_root: Path | None = None, database_url: str | None = None) -> None:
         self._catalog_runtime = CatalogRuntime(
@@ -459,6 +461,7 @@ class ComponentCatalogDomainService:
         self._component_read_models = CatalogComponentReadModels(self._revision_kernel)
         self._component_queries = CatalogComponentQueries(self._component_read_models)
         self._project_import_sessions = CatalogProjectImportSessions()
+        self._project_import_matching = CatalogProjectImportMatching()
 
     def _runtime_for_compat(self) -> CatalogRuntime:
         """Lazily support legacy ``__new__``-constructed test doubles."""
@@ -1451,145 +1454,27 @@ class ComponentCatalogDomainService:
             for row in rows
         ]
 
-    def _record_component_usage(
-        self,
-        conn: Any,
-        *,
-        component_id: str,
-        provenance: list[dict[str, Any]],
-        observed_at: str,
-        source: str = "semantic_scan",
-    ) -> int:
-        grouped: dict[tuple[str, str], list[dict[str, Any]]] = {}
-        for raw_source in provenance:
-            project_id = str(raw_source.get("projectId") or "")
-            source_revision = str(raw_source.get("sourceRevision") or "")
-            if not project_id:
-                continue
-            detail = {
-                str(key): value
-                for key, value in raw_source.items()
-                if value not in (None, "", [], {})
-            }
-            grouped.setdefault((project_id, source_revision), []).append(detail)
-
-        for (project_id, source_revision), details in grouped.items():
-            conn.execute(
-                """
-                UPDATE component_usage
-                SET is_current = 0, last_seen_at = %s
-                WHERE component_id = %s AND project_id = %s AND source_revision <> %s AND is_current = 1
-                """,
-                (observed_at, component_id, project_id, source_revision),
-            )
-            references = sorted(
-                {
-                    str(detail.get("reference") or "")
-                    for detail in details
-                    if str(detail.get("reference") or "")
-                }
-            )
-            conn.execute(
-                """
-                INSERT INTO component_usage (
-                    id, component_id, project_id, source_revision, references_json, details_json,
-                    source, is_current, first_seen_at, last_seen_at
-                ) VALUES (%s, %s, %s, %s, %s, %s, %s, 1, %s, %s)
-                ON CONFLICT(component_id, project_id, source_revision)
-                DO UPDATE SET references_json = excluded.references_json,
-                              details_json = excluded.details_json,
-                              source = excluded.source,
-                              is_current = 1,
-                              last_seen_at = excluded.last_seen_at
-                """,
-                (
-                    str(uuid.uuid4()),
-                    component_id,
-                    project_id,
-                    source_revision,
-                    json.dumps(references, separators=(",", ":")),
-                    json.dumps(details, sort_keys=True, separators=(",", ":")),
-                    source,
-                    observed_at,
-                    observed_at,
-                ),
-            )
-        return len(grouped)
-
     def index_project_component_usage(self, proposals: list[dict[str, Any]]) -> dict[str, int]:
         """Index where-used observations even when no import proposal is accepted."""
         self.initialize()
-        matched_components: set[str] = set()
-        observations = 0
         now = _utc_now_iso()
         with self._connect() as conn:
-            for proposal in proposals:
-                metadata = dict(proposal.get("metadata") or {})
-                manufacturer = str(metadata.get("manufacturer") or "").strip()
-                mpn = str(metadata.get("manufacturer_part_number") or metadata.get("mpn") or "").strip()
-                if not manufacturer or not mpn:
-                    continue
-                component = conn.execute(
-                    """
-                    SELECT component.id
-                    FROM components component
-                    JOIN component_revisions revision ON revision.id = component.current_revision_id
-                    WHERE component.is_active = 1
-                      AND lower(revision.manufacturer) = lower(%s)
-                      AND lower(revision.mpn) = lower(%s)
-                    ORDER BY component.created_at
-                    LIMIT 1
-                    """,
-                    (manufacturer, mpn),
-                ).fetchone()
-                if not component:
-                    continue
-                component_id = str(component["id"])
-                matched_components.add(component_id)
-                observations += self._record_component_usage(
-                    conn,
-                    component_id=component_id,
-                    provenance=list(proposal.get("provenance") or []),
-                    observed_at=now,
-                    source="semantic_scan",
-                )
+            result = self._project_import_matching.index_project_component_usage(
+                conn,
+                proposals,
+                observed_at=now,
+            )
             conn.commit()
-        return {"matched_components": len(matched_components), "observations": observations}
+        return result
 
     def match_component_identities(self, identities: list[dict[str, str]]) -> dict[str, dict[str, Any]]:
         """Resolve manufacturer/MPN identities in one catalog query for import preflight."""
-        requested = {
-            (str(item.get("manufacturer") or "").strip().casefold(), str(item.get("mpn") or "").strip().casefold())
-            for item in identities
-            if str(item.get("manufacturer") or "").strip() and str(item.get("mpn") or "").strip()
-        }
+        requested = self._project_import_matching.normalize_identity_requests(identities)
         if not requested:
             return {}
         self.initialize()
         with self._connect() as conn:
-            rows = conn.execute(
-                """
-                SELECT component.id, revision.name, revision.manufacturer, revision.mpn,
-                       revision.id AS revision_id, revision.version
-                FROM components component
-                JOIN component_revisions revision ON revision.id = component.current_revision_id
-                WHERE component.is_active = 1
-                """
-            ).fetchall()
-        matches: dict[str, dict[str, Any]] = {}
-        for row in rows:
-            key = (str(row["manufacturer"] or "").strip().casefold(), str(row["mpn"] or "").strip().casefold())
-            if key not in requested:
-                continue
-            matches["\0".join(key)] = {
-                "component_id": str(row["id"]),
-                "revision_id": str(row["revision_id"]),
-                "version": int(row["version"] or 0),
-                "name": str(row["name"] or row["mpn"] or ""),
-                "manufacturer": str(row["manufacturer"] or ""),
-                "manufacturer_part_number": str(row["mpn"] or ""),
-            }
-        return matches
+            return self._project_import_matching.match_component_identities(conn, requested)
 
     def _revision_matches_import(
         self,
@@ -1863,7 +1748,7 @@ class ComponentCatalogDomainService:
                         "provenance": provenance,
                     },
                 )
-            self._record_component_usage(
+            self._project_import_matching.record_component_usage(
                 conn,
                 component_id=component_id,
                 provenance=provenance,
